python/requirements.txt
tweepy
python-dotenv
cryptocode